requirements.txt
python-dotenv
python-kis
openai
pandas
requests